requirements-minimal.txt
# Minimal requirements to run the API
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
python-multipart==0.0.20

# Database
elasticsearch==8.17.0

# AI/ML
groq==0.14.0

# Security & Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4

# Data Processing
pydantic==2.10.5
pydantic-settings==2.7.0

# PII Detection & Redaction
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Monitoring & Logging
structlog==24.4.0

# Development
python-dotenv==1.0.1